蜂蜜层bug

+++ src/main/java/net/Gmaj7/funny_world/daiBlocks/custom/HoneyFloorBlock.java
@@ -1,6 +1,8 @@
 package net.Gmaj7.funny_world.daiBlocks.custom;
 
 import com.mojang.serialization.MapCodec;
+import net.Gmaj7.funny_world.daiBlocks.blockEntity.HoneyFloorBlockEntity;
+import net.Gmaj7.funny_world.daiInit.daiDataComponentTypes;
 import net.Gmaj7.funny_world.daiInit.daiHoneyEffects;
 import net.Gmaj7.funny_world.daiInit.daiUniqueData.daiUniqueDataGet;
 import net.minecraft.core.BlockPos;
@@ -17,16 +19,20 @@ import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.LevelReader;
+import net.minecraft.world.level.block.BaseEntityBlock;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.RenderShape;
+import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 
-public class HoneyFloorBlock extends Block {
+public class HoneyFloorBlock extends BaseEntityBlock {
     public static final MapCodec<HoneyFloorBlock> CODEC = simpleCodec(HoneyFloorBlock::new);
     protected static final VoxelShape SHAPE = Block.box(0.0, 0.0, 0.0, 16.0, 1.0, 16.0);
     public HoneyFloorBlock(Properties properties) {
@@ -56,27 +62,39 @@ public class HoneyFloorBlock extends Block {
             double d1 = 0.4 + d0 * 0.2;
             entity.setDeltaMovement(entity.getDeltaMovement().multiply(d1, 1.0, d1));
         }
-        List<daiHoneyEffects.Entry> list = ((daiUniqueDataGet)state).getHoneyAbsorbEffect().getEffect();
-        if(list != null && !list.isEmpty() && entity instanceof LivingEntity livingEntity){
-            for (daiHoneyEffects.Entry entry : list){
-                livingEntity.addEffect(new MobEffectInstance(entry.effect(), entry.duration(), entry.effectLevel()));
-            }
+        HoneyFloorBlockEntity honeyFloorBlockEntity = level.getBlockEntity(pos) instanceof HoneyFloorBlockEntity ? (HoneyFloorBlockEntity) level.getBlockEntity(pos) : new HoneyFloorBlockEntity(pos, state);
+        if(honeyFloorBlockEntity.getHoney_bottle().has(daiDataComponentTypes.HONEY_EFFECTS) && entity instanceof LivingEntity target){
+            List<daiHoneyEffects.Entry> list = honeyFloorBlockEntity.getHoney_bottle().get(daiDataComponentTypes.HONEY_EFFECTS).effects();
+            for (daiHoneyEffects.Entry entry : list)
+                target.addEffect(new MobEffectInstance(entry.effect(), entry.duration(), entry.effectLevel()));
         }
         super.entityInside(state, level, pos, entity);
     }
 
     @Override
     protected ItemInteractionResult useItemOn(ItemStack stack, BlockState state, Level level, BlockPos pos, Player player, InteractionHand hand, BlockHitResult hitResult) {
-        if(stack.is(Items.GLASS_BOTTLE)){
+        BlockEntity blockEntity = level.getBlockEntity(pos);
+        if(stack.is(Items.GLASS_BOTTLE) && blockEntity instanceof HoneyFloorBlockEntity honeyFloorBlockEntity){
             if(!player.isCreative())
                 stack.shrink(1);
-            ItemStack itemStack = new ItemStack(Items.HONEY_BOTTLE);
+            ItemStack itemStack = honeyFloorBlockEntity.getHoney_bottle();
             if (!player.getInventory().add(itemStack)) {
                 player.drop(itemStack, false);
             }
-            level.destroyBlock(pos, false);
+                    level.destroyBlock(pos, false);
             player.swing(hand);
         }
         return super.useItemOn(stack, state, level, pos, player, hand, hitResult);
+    }
+
+    @Nullable
+    @Override
+    public BlockEntity newBlockEntity(BlockPos blockPos, BlockState blockState) {
+        return new HoneyFloorBlockEntity(blockPos, blockState);
+    }
+
+    @Override
+    protected RenderShape getRenderShape(BlockState state) {
+        return RenderShape.MODEL;
     }
 }
